+++ src/WeihanLi.Common/Event/EventBusExtensions.cs
@@ -1,0 +1,6 @@
+﻿namespace WeihanLi.Common.Event
+{
+    public static class EventBusExtensions
+    {
+    }
+}