+++ src/WeihanLi.Common/Helpers/TotpHelper.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using WeihanLi.Common.Otp;
 
 namespace WeihanLi.Common.Helpers
 {